+++ tests/ruby/misc_spec.rb
@@ -1,10 +1,44 @@
 # frozen_string_literal: true
 require_relative 'spec_helper'
 
-
 describe "Miscellaneous" do
   let(:proxies) { Helpers::Pgcat.single_shard_setup("sharded_db", 5) }
   after { proxies.all&.map(&:shutdown) }
+
+  describe "Extended Protocol handling" do
+    it "does not send packets that client does not expect during extended protocol sequence" do
+      new_configs = proxies.main.current_config
+
+      new_configs["general"]["connect_timeout"] = 500
+      new_configs["general"]["ban_time"] = 1
+      new_configs["general"]["shutdown_timeout"] = 1
+      new_configs["pools"]["sharded_db"]["users"]["0"]["pool_size"] = 1
+
+      proxies.main.update_config(new_configs)
+      proxies.main.reload_config
+
+      25.times do
+        Thread.new do
+          conn = PG::connect(proxies.main.connection_string("sharded_db", "sharding_user"))
+          conn.async_exec("SELECT pg_sleep(5)") rescue PG::SystemError
+        ensure
+          conn&.close
+        end
+      end
+
+      sleep(0.5)
+      conn_under_test = PG::connect(proxies.main.connection_string("sharded_db", "sharding_user"))
+      stdout, stderr = with_captured_stdout_stderr do
+        15.times do |i|
+          conn_under_test.async_exec("SELECT 1") rescue PG::SystemError
+          conn_under_test.exec_params("SELECT #{i} + $1", [i]) rescue PG::SystemError
+          sleep 1
+        end
+      end
+
+      raise StandardError, "Libpq got unexpected messages while idle" if stderr.include?("arrived from server while idle")
+    end
+  end
 
   describe "Pool recycling after config reload" do
     let(:proxies) { Helpers::Pgcat.three_shard_setup("sharded_db", 5) }
